@@ -1,8 +1,8 @@
 {
   "_from": "basic-auth@~2.0.0",
-  "_id": "basic-auth@2.0.0",
+  "_id": "basic-auth@2.0.1",
   "_inBundle": false,
-  "_integrity": "sha1-AV2z81PgLlY3d1X5YnQuiYHnu7o=",
+  "_integrity": "sha512-NF+epuEdnUYVlGuhaxbbq+dvJttwLnGY+YixlXlME5KpQ5W3CnXA5cVTneY3SPbPDRkcjMbifrwmFYcClgOZeg==",
   "_location": "/basic-auth",
   "_phantomChildren": {},
   "_requested": {
@@ -18,29 +18,29 @@
   "_requiredBy": [
     "/morgan"
   ],
-  "_resolved": "https://registry.npmjs.org/basic-auth/-/basic-auth-2.0.0.tgz",
-  "_shasum": "015db3f353e02e56377755f962742e8981e7bbba",
+  "_resolved": "https://registry.npmjs.org/basic-auth/-/basic-auth-2.0.1.tgz",
+  "_shasum": "b998279bf47ce38344b4f3cf916d4679bbf51e3a",
   "_spec": "basic-auth@~2.0.0",
-  "_where": "/Users/joaohenrique/Repositories/freeCodeCamp/my_awesome_project/api/node_modules/morgan",
+  "_where": "C:\\Users\\track\\Documents\\mini-full-stack-project\\api\\node_modules\\morgan",
   "bugs": {
     "url": "https://github.com/jshttp/basic-auth/issues"
   },
   "bundleDependencies": false,
   "dependencies": {
-    "safe-buffer": "5.1.1"
+    "safe-buffer": "5.1.2"
   },
   "deprecated": false,
   "description": "node.js basic auth parser",
   "devDependencies": {
-    "eslint": "3.19.0",
-    "eslint-config-standard": "10.2.1",
-    "eslint-plugin-import": "2.7.0",
+    "eslint": "5.6.0",
+    "eslint-config-standard": "12.0.0",
+    "eslint-plugin-import": "2.14.0",
     "eslint-plugin-markdown": "1.0.0-beta.6",
-    "eslint-plugin-node": "5.1.1",
-    "eslint-plugin-promise": "3.5.0",
-    "eslint-plugin-standard": "3.0.1",
+    "eslint-plugin-node": "7.0.1",
+    "eslint-plugin-promise": "4.0.1",
+    "eslint-plugin-standard": "4.0.0",
     "istanbul": "0.4.5",
-    "mocha": "2.5.3"
+    "mocha": "5.2.0"
   },
   "engines": {
     "node": ">= 0.8"
@@ -69,5 +69,5 @@
     "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter dot --check-leaks test/",
     "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --reporter spec --check-leaks test/"
   },
-  "version": "2.0.0"
+  "version": "2.0.1"
 }
