@@ -1,27 +1,28 @@
 {
-  "_from": "morgan@~1.9.0",
-  "_id": "morgan@1.9.0",
+  "_from": "morgan@1.9.1",
+  "_id": "morgan@1.9.1",
   "_inBundle": false,
-  "_integrity": "sha1-0B+mxlhZt2/PMbPLU6OCGjEdgFE=",
+  "_integrity": "sha512-HQStPIV4y3afTiCYVxirakhlCfGkI161c76kKFca7Fk1JusM//Qeo1ej2XaMniiNeaZklMVrh3vTtIzpzwbpmA==",
   "_location": "/morgan",
   "_phantomChildren": {},
   "_requested": {
-    "type": "range",
+    "type": "version",
     "registry": true,
-    "raw": "morgan@~1.9.0",
+    "raw": "morgan@1.9.1",
     "name": "morgan",
     "escapedName": "morgan",
-    "rawSpec": "~1.9.0",
+    "rawSpec": "1.9.1",
     "saveSpec": null,
-    "fetchSpec": "~1.9.0"
+    "fetchSpec": "1.9.1"
   },
   "_requiredBy": [
+    "#USER",
     "/"
   ],
-  "_resolved": "https://registry.npmjs.org/morgan/-/morgan-1.9.0.tgz",
-  "_shasum": "d01fa6c65859b76fcf31b3cb53a3821a311d8051",
-  "_spec": "morgan@~1.9.0",
-  "_where": "/Users/joaohenrique/Repositories/freeCodeCamp/my_awesome_project/api",
+  "_resolved": "https://registry.npmjs.org/morgan/-/morgan-1.9.1.tgz",
+  "_shasum": "0a8d16734a1d9afbc824b99df87e738e58e2da59",
+  "_spec": "morgan@1.9.1",
+  "_where": "C:\\Users\\track\\Documents\\mini-full-stack-project\\api",
   "bugs": {
     "url": "https://github.com/expressjs/morgan/issues"
   },
@@ -40,20 +41,20 @@
   "dependencies": {
     "basic-auth": "~2.0.0",
     "debug": "2.6.9",
-    "depd": "~1.1.1",
+    "depd": "~1.1.2",
     "on-finished": "~2.3.0",
     "on-headers": "~1.0.1"
   },
   "deprecated": false,
   "description": "HTTP request logger middleware for node.js",
   "devDependencies": {
-    "eslint": "3.19.0",
-    "eslint-config-standard": "10.2.1",
-    "eslint-plugin-import": "2.7.0",
+    "eslint": "5.5.0",
+    "eslint-config-standard": "12.0.0",
+    "eslint-plugin-import": "2.14.0",
     "eslint-plugin-markdown": "1.0.0-beta.6",
-    "eslint-plugin-node": "5.1.1",
-    "eslint-plugin-promise": "3.5.0",
-    "eslint-plugin-standard": "3.0.1",
+    "eslint-plugin-node": "7.0.1",
+    "eslint-plugin-promise": "4.0.1",
+    "eslint-plugin-standard": "4.0.0",
     "istanbul": "0.4.5",
     "mocha": "2.5.3",
     "split": "1.0.1",
@@ -87,5 +88,5 @@
     "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --check-leaks --reporter dot",
     "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --check-leaks --reporter spec"
   },
-  "version": "1.9.0"
+  "version": "1.9.1"
 }
